Add skipLibCheck option to tsconfig.json
I'm not entirely clear on the side-effects of doing this, it seems like the wrong solution. The problem was that the mdx-bundler packaage had incorrect types and was causing errors in the typecheck, even though that wasn't my code. After investigating other approaches, like making my own .d.ts file with a  statement, using skipLibCheck was the only one that worked.